Pistol shooting demo
Holding/clicking left click shoots a pistol bullet, only allows a shot
to be fired every 0.5s

+++ Assets/Scripts/Weapon.cs
@@ -5,7 +5,7 @@ public interface IWeapon
 
     void Fire();
 
-    double FireDelay
+    float FireDelay
     {
         get; set;
     }
